Adds shared mock web storage
